Data generator and CSV creator working
