+++ db/schemas/schemaMongo.js
@@ -1,4 +1,4 @@
-const db = require('./index.js');
+const db = require('../index.js');
 const mongoose = require('mongoose');
 
 const productSchema = mongoose.Schema({
