+++ db/seed/seedPSQL.sh
@@ -1,0 +1,1 @@
+psql postgres
